Data.Function.Class

Description

Documentation

class Function f where

An instance of Function is a type that represents functions. Function can be evaluated at points in their Domain.

Examples: Polynomials, trigonometric polynomials, piecewise polynomials, …

Associated Types

type Domain f

The domain of definition of the function.

type Codomain f

The codomain of a function is the set of potential function values, i.e. function values never lie outside this set.

In contrast, the set of actual function values is called the image and is typically a strict subset of the codomain.

Methods

eval :: f -> Domain f -> Codomain f

Evaluate a function at a point in its Domain.