Numeric.Measure.Discrete

Description

Documentation

data Discrete a

A discrete, finite signed measure on the number line.

fromMap :: (Ord a, Num a) => Map a a -> Discrete a

Construct a discrete measure from a collection of points and their measures.

toMap :: Num a => Discrete a -> Map a a

Decompose the discrete measure into a collection of points and their measures.

dirac :: (Ord a, Num a) => a -> a -> Discrete a

A Dirac measure at the given point x.

total (dirac x m) = m 

total :: Num a => Discrete a -> a

The total of the measure applied to the set of real numbers.

distribution :: (Ord a, Num a) => Discrete a -> a -> a

eval (distribution m) x is the measure of the interval $(-∞, x]$.

This is known as the distribution function).

add :: (Ord a, Num a) => Discrete a -> Discrete a -> Discrete a

Add two measures.

total (add mx my) = total mx + total my

convolve :: (Ord a, Num a) => Discrete a -> Discrete a -> Discrete a

Additive convolution of two measures.

convolve (dirac x wx) (dirac y wy) = dirac (x + y) (wx * wy)
convolve (add mx my) mz = add (convolve mx mz) (convolve my mz)
convolve mx (add my mz) = add (convolve mx my) (convolve mx mz)
total (convolve mx my) = total mx * total my