Numeric.Measure.Finite.Mixed

Description

Type

data Measure a

A finite signed measure on the number line.

zero :: Num a => Measure a

The measure that assigns 0 to every set.

dirac :: (Ord a, Num a) => a -> Measure a

A Dirac measure at the given point x.

total (dirac x) = 1

uniform :: (Ord a, Num a, Fractional a) => a -> a -> Measure a

The probability measure of a uniform probability distribution in the interval \( [x,y) \).

total (uniform x y) = 1

distribution :: (Ord a, Num a) => Measure a -> Piecewise (Poly a)

eval (distribution m) x is the measure of the interval \( (-∞, x] \).

This is known as the distribution function.

fromDistribution :: (Ord a, Num a) => Piecewise (Poly a) -> Maybe (Measure a)

Construct a signed measure from its distribution function.

Return Nothing if the measure is not finite, that is if the last piece of the piecewise function is not constant.

Observations

total :: (Ord a, Num a) => Measure a -> a

The total of the measure applied to the set of real numbers.

support :: (Ord a, Num a) => Measure a -> Maybe (a, a)

The support is the smallest closed, contiguous interval \( [x,y] \) outside of which the measure is zero.

Returns Nothing if the interval is empty.

isPositive :: (Ord a, Num a, Fractional a) => Measure a -> Bool

Check whether a signed measure is positive.

A signed measure is positive if the measure of any set is nonnegative. In other words a positive signed measure is just a measure in the ordinary sense.

This test is nontrivial, as we have to check that the distribution function is monotonically increasing.

integrate :: (Ord a, Num a, Fractional a) => Poly a -> Measure a -> a

Integrate a polynomial f with respect to the given measure m, \( \int f(x) dm(x) \).

Operations, numerical

add :: (Ord a, Num a) => Measure a -> Measure a -> Measure a

Add two measures.

total (add mx my) = total mx + total my

scale :: (Ord a, Num a) => a -> Measure a -> Measure a

Scale a measure by a constant.

total (scale a mx) = a * total mx

translate :: (Ord a, Num a, Fractional a) => a -> Measure a -> Measure a

Translate a measure along the number line.

eval (distribution (translate y m)) x
   = eval (distribution m) (x - y)

convolve :: (Ord a, Num a, Fractional a) => Measure a -> Measure a -> Measure a

Additive convolution of two measures.

Properties:

convolve (dirac x) (dirac y) = dirac (x + y)

convolve mx my               =  convolve my mx
convolve (add mx my) mz      =  add (convolve mx mz) (convolve my mz)
translate z (convolve mx my) =  convolve (translate z mx) my
total (convolve mx my)       =  total mx * total my