Numeric.Measure.Probability

Description

Type

data Prob a

A probability measure on the number line.

A probability measure is a Measure whose total is 1.

dirac :: (Ord a, Num a) => a -> Prob a

A Dirac measure at the given point x.

dirac x is the probability distribution where x occurs with certainty.

uniform :: (Ord a, Num a, Fractional a) => a -> a -> Prob a

The uniform probability distribution on the interval \( [x,y) \).

distribution :: (Ord a, Num a) => Prob a -> Piecewise (Poly a)

eval (distribution m) x is the probability of picking a number <= x.

This is known as the cumulative distribution function.

fromDistribution :: (Ord a, Num a, Fractional a) => Piecewise (Poly a) -> Maybe (Prob a)

Construct a probability distribution from its cumulative distribution function.

Return Nothing if * the cumulative distribution function is not monotonicall increasing * the last piece of the piecewise function is not a constant equal to 1.

measure :: (Ord a, Num a) => Prob a -> Measure a

View the probability measure as a Measure.

fromMeasure :: (Ord a, Num a, Fractional a) => Measure a -> Maybe (Prob a)

View a Measure as a probability distribution.

The measure m must be positive, with total weight 1, that is

isPositive m == True
total m == 1

These preconditions are checked and the function returns Nothing if they fail.

unsafeFromMeasure :: Measure a -> Prob a

View a Measure as a probability distribution.

Variant of fromMeasure where the precondition are not checked!

Observations

support :: (Ord a, Num a) => Prob a -> Maybe (a, a)

The support is the smallest closed, contiguous interval \( [x,y] \) outside of which the probability is zero.

Returns Nothing if the interval is empty.

expectation :: (Ord a, Num a, Fractional a) => Poly a -> Prob a -> a

Compute the expected value of a polynomial f with respect to the given probability distribution, \( E[f(X)] \).

moments :: (Ord a, Num a, Fractional a) => Prob a -> Moments a

Compute the first four commonly used moments of a probability distribution.

Operations, numerical

choice :: (Ord a, Num a, Fractional a) => a -> Prob a -> Prob a -> Prob a

Left-biased random choice.

choice p is a probability distribution where events from the left argument are chosen with probablity p and events from the right argument are chosen with probability (1-p).

eval (distribution (choice p mx my)) z
   = p * eval (distribution mx) z + (1-p) * eval (distribution my) z

translate :: (Ord a, Num a, Fractional a) => a -> Prob a -> Prob a

Translate a probability distribution along the number line.

eval (distribution (translate y m)) x
   = eval (distribution m) (x - y)

convolve :: (Ord a, Num a, Fractional a) => Prob a -> Prob a -> Prob a

Additive convolution of two probability measures.

Properties:

convolve (dirac x) (dirac y) = dirac (x + y)

convolve mx my               =  convolve my mx
translate z (convolve mx my) =  convolve (translate z mx) my