Numeric.Polynomial.Simple

Description

Polynomials and computations with them.

Basic operations

data Poly a

Polynomial with coefficients in a.

eval :: Num a => Poly a -> a -> a

Evaluate a polynomial at a point.

Uses Horner's method to minimise the number of multiplications.

a0 + a1·x + a2·x^2 + ... + a{n-1}·x^{n-1} + an·x^n
  = a0 + x·(a1 + x·(a2 + x·(… + x·(a{n-1} + x·an)) ))

degree :: (Eq a, Num a) => Poly a -> Int

Degree of a polynomial.

The degree of a constant polynomial is 0, but the degree of the zero polynomial is -1 for Euclidean division.

constant :: a -> Poly a

The constant polynomial.

eval (constant a) = const a

zero :: Num a => Poly a

The zero polynomial.

monomial :: (Eq a, Num a) => Int -> a -> Poly a

monomial n a is the polynomial a * x^n.

fromCoefficients :: Num a => [a] -> Poly a

Construct a polynomial a0 + a1·x + … from its list of coefficients [a0, a1, …].

toCoefficients :: Poly a -> [a]

List the coefficients [a0, a1, …] of a polynomial a0 + a1·x + ….

scale :: Num a => a -> Poly a -> Poly a

Scale a polynomial by a scalar. More efficient than multiplying by a constant polynomial.

eval (scale a p) x = a * eval p x

scaleX :: (Eq a, Num a) => Poly a -> Poly a

Multiply the polynomial by the unknown x.

eval (scaleX p) x = x * eval p x
degree (scaleX p) = 1 + degree p  if  degree p >= 0

Advanced operations

Convenience

display :: (Ord a, Eq a, Num a) => Poly a -> (a, a) -> a -> [(a, a)]

Return a list of pairs (x, eval p x) from the graph of the polynomial. The values x are from the range (l, u) with uniform spacing s.

Specifically,

map fst (display p (l, u) s) = [l, l+s, l + 2·s, … , u']

where u' is the largest number of the form u' = l + s·k, k natural, that still satisfies u' < l.

lineFromTo :: (Eq a, Fractional a) => (a, a) -> (a, a) -> Poly a

Linear polymonial connecting the points (x1, y1) and (x2, y2), assuming that x1 ≠ x2.

let p = lineFromTo (x1, y1) (x2, y2)

degree p <= 1
eval p x1 = y1
eval p x2 = y2

Algebraic

translate :: (Fractional a, Eq a, Num a) => a -> Poly a -> Poly a

Translate the argument of a polynomial by summing binomial expansions.

eval (translate y p) x = eval p (x - y)

integrate :: (Eq a, Fractional a) => Poly a -> Poly a

Indefinite integral of a polynomial with constant term zero.

The integral of x^n is 1/(n+1)·x^(n+1).

eval (integrate p) 0 = 0
integrate (differentiate p) = p - constant (eval p 0)

differentiate :: Num a => Poly a -> Poly a

Differentiate a polynomial.

We have dx^n/dx = n·x^(n-1).

differentiate (integrate p) = p
differentiate (p * q) = (differentiate p) * q + p * (differentiate q)

convolve :: (Fractional a, Eq a, Ord a) => (a, a, Poly a) -> (a, a, Poly a) -> [(a, Poly a)]

Convolution of two polynomials defined on bounded intervals. Produces three contiguous pieces as a result.

Numerical

compareToZero :: (Fractional a, Eq a, Ord a) => (a, a, Poly a) -> Maybe Ordering

Measure whether or not a polynomial is consistently above or below zero, or equals zero.

Need to consider special cases where there is a root at a boundary point.

countRoots :: (Fractional a, Eq a, Ord a) => (a, a, Poly a) -> Int

We use Sturm's Theorem to count the number of roots of a polynomial in a given interval.

(See https://en.wikipedia.org/wiki/Sturm%27s_theorem) Starting from polynomial p, construct the Sturm sequence p0, p1, . . ., where: p0 = p p1 = p′ pi+1 = −rem(pi−1, pi) for i > 1 where p′ is the derivative of p and rem(p, q) is the remainder of the Euclidian division of p by q. The length of this sequence is at most the degree of p. We define V(x) to be the number of sign variations in the sequence of numbers p0(x), p1(x), . . .. Sturm’s theorem states that, if p is a square-free polynomial (one without repeated roots), then R(l,r](p) = V (l) − V (r). This extends to non-square-free polynomials provided neither l nor r is a multiple root of p (a circumstance we shall ignore)

We start from the tuple that emerges from disagregation.

findRoot :: (Fractional a, Eq a, Num a, Ord a) => a -> (a, a) -> Poly a -> Maybe a

Find the root of a polynomial in a given interval, assuming that there is exactly one root in the given interval. This precondition has to be checked through other means, e.g. countRoots.

We find the root by repeatedly halving the interval in which the root must lie. Constant and linear polynomials, degree p <= 1, are treated as special cases.