PWPs.IRVs

Description

We build IRVs as an abstract datatype on top of piecewise polynomials and deltas. IRVs may be represented as PDFs or CDFs. They are polymorphic in a constrained numeric type.

We provide functions to construct IRVs as uniform distributions on an interval; a delta at a point; and a CDF from a list of point-pairs.

We provide functions to turn an IRV into a set of point-pairs representing either the PDF or CDF.

We define operators for convolution, first-to-finish, all-to-finish and weighted choice. We also provide an operation to extract the probability mass (<= 1) and the support of the distribution.

We provide an implementation of the partial order on IRVs.

We enable extraction of centiles and computation of moments.

Note that we can consistently represent bottom using polynomials and top by including deltas.

Documentation

data IRV a

makePDF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => IRV a -> DistD a

Force an IRV into a PDF by differentiating if necessary

makeCDF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> DistP a

Force an IRV into a CDF by integrating if necessary

constructUniform :: (Fractional a, Ord a, Num a, Enum a, Eq a) => a -> IRV a

Construct a PDF with uniform probability from 0 to the given value

constructDelta :: (Fractional a, Ord a, Num a, Enum a, Eq a) => a -> IRV a

Construct a PDF that is a delta function at the given value

zeroPDF :: (Fractional a, Ord a, Num a, Enum a, Eq a) => IRV a

PDF with zero value everywhere

constructCDF :: (Fractional a, Ord a, Num a, Enum a, Eq a) => [(a, a)] -> IRV a

Construct a CDF from a list of values, treating each new value as a step up from the one before, assuming we start at 0 | First interval is a zero polynomial: subsequent intervals have a constant polynomial.

constructLinearCDF :: (Fractional a, Ord a, Num a, Enum a, Eq a) => [(a, a)] -> IRV a

Construct a CDF from a list of values, interpolating linearly between each pair of points

constructGeneralCDF :: (Fractional a, Ord a, Num a, Enum a, Eq a) => [(a, a)] -> IRV a

Construct a CDF from a list of (time, probability) pairs, which start at (0,0) and are monotonic. If the next time is strictly greater than the previous one, construct a linear interpolation between the two probabilities. If the next time is the same as the previous one, construct a Heaviside step between the two probabilities.

firstToFinish :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> IRV a -> IRV a

multiFtF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => [IRV a] -> IRV a

Compute the first-to-finsh of a list of IRVs by multiplying inverse CDFs then inverting If there is nothing finish the result is bottom

allToFinish :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> IRV a -> IRV a

multiAtF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => [IRV a] -> IRV a

Compute the last-to-finsh of a list of IRVs: if we have nothing to wait for the result is top

probChoice :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => a -> IRV a -> IRV a -> IRV a

The probability is for choosing the left branch. We can do this on either PDFs or CDFs; if we have CDFs deliver a CDF, if we have both PDFs or one of each, deliver a PDF.

multiWeightedChoice :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => [(a, IRV a)] -> IRV a

(<+>) :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => IRV a -> IRV a -> IRV a infix 7

To convolve, force into PDFs and then invoke piecewise convolution

probMass :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> a

Just extract the final value of the CDF

asDiscreteCDF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> Int -> [Either (a, a) [(a, a)]]

Return a sequence of (Left) (time, step base) (the lower value of the jump at that point) or (Right) a sequence of (Time, Probability). The sequence is monotonically increasing in Time.

asDiscretePDF :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => IRV a -> Int -> [Either (a, a) [(a, a)]]

Return a sequence of (Left) Impulse Probablity mass (equivalent to the integral of the Heaviside function at that point) or (Right) a sequence of Time and Probability Density. The sequence is monotonically increasing in Time.

compareIRVs :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a)) => IRV a -> IRV a -> Maybe Ordering

If the two IRVs are partially ordered, return an ordering, otherwise return Nothing.

support :: (Eq a, Fractional a) => IRV a -> (a, a)

return the first and last basepoints for which the value is significant

top :: (Ord a, Num a) => IRV a

bottom :: (Ord a, Num a) => IRV a

centiles :: (Fractional a, Ord a, Num a, Enum a, Eq a, Evaluable a (Poly a), Integrable (DistD a) (DistP a)) => [a] -> IRV a -> [Maybe a]

Given an ordered list of probabiity values, return the time at which each value is reached; if it is never reached, return Nothing in that position. Where there is a step up in the CDF that includes the centile probability, we report the time for which the probability is exceeded. Centiles are not computed exactly, but to a given precision, currently set to 1/1000th of the support of the IRV.

cumulativeMass :: (Ord a, Enum a, Num a, Fractional a, Evaluable a (DistP a), Integrable (DistD a) (DistP a)) => IRV a -> a -> a

shiftIRV :: (Fractional a, Ord a, Num a, Enum a, Eq a, Differentiable (DistP a) (DistD a)) => a -> IRV a -> IRV a

Make a delta and convolve with it

data Moments a

Constructors

Moments

Fields

tangibleMass :: a

mean :: a

variance :: a

skew :: a

kurtosis :: a

moments :: (Fractional a, Ord a, Num a, Enum a, Eq a, Integrable (DistD a) (DistP a), Differentiable (DistP a) (DistD a), Evaluable a (DistD a), Evaluable a (DistP a)) => IRV a -> Moments a

Compute the first five moments of a given distribution

complexityOfIRV :: (Fractional a, Ord a, Num a, Enum a, Differentiable (DistP a) (DistD a), ComplexityMeasureable (PolyDelta a)) => IRV a -> Int